Return correct exit codes.
These changes resolve a bash error when trying to exit with a string value.
Fossa commands were run and their output was stored in a variable. That
variable was then passed to exit. Exit expects a numeric value between 0
and 255 and not a string so it gave this error:

exit: Test: numeric argument required

This change stores the return code from the fossa run in the variable that
later passed to exit.